Update sns-sqs-as-subscriber.md
On line #3 I recommend changing "Both" to "These". When "Both" is used it implies that each service offers different benefits, when the purpose of the sentence is to explain that the services are different.   Another option would be to use the word "unique" as in "Both services provide unique benefits for developers". 
It's a minor change, I don't think the sentence as it exists now will be misconstrued by the reader but it is more appropriate to use one of the modifications I suggested. 
Thanks for taking the time to review my proposal.